+++ script/core/diagnostics/cast-local-type.lua
@@ -22,9 +22,17 @@ return function (uri, callback)
             return
         end
         for _, ref in ipairs(loc.ref) do
-            if ref.type == 'setlocal' then
+            if ref.type == 'setlocal' and ref.value then
                 await.delay()
                 local refNode = vm.compileNode(ref)
+                local value = ref.value
+
+                if value.type == 'getfield'
+                or value.type == 'getindex' then
+                    -- 由于无法对字段进行类型收窄，
+                    -- 因此将假值移除再进行检查
+                    refNode = refNode:copy():setTruthy()
+                end
 
                 if not vm.canCastType(uri, locNode, refNode) then
                     callback {